+++ CHANGELOG.md
@@ -5,6 +5,7 @@
  * Added requirement for min Ruby Gems version in line with Rails 7.2
  * Fixed country_select compatibility by switching from v1 to v2 syntax ([#1382](https://github.com/formtastic/formtastic/pull/1382))
  * Fixed that passing caller to Deprecation#warn is deprecated ([#1384](https://github.com/formtastic/formtastic/pull/1384))
+ * Added `lable_html` option for overriding & adding label HTML attributes ([#1379](https://github.com/formtastic/formtastic/pull/1379))
 
 ## 5.0.0